refactor: loop <- teamLoop(mainGoal, subGoal) (#41)
- teamLoop 하위 도메인 이동, issuiTemplate(Refactor) label 추가
