+++ .github/ISSUE_TEMPLATE/리팩토링.md
@@ -1,8 +1,8 @@
 ---
 name: 리팩토링
-about: 리팩토링 진행
-title: "[REFACT] "
-labels: ''
+about: '리팩토링 진행'
+title: "[REFACT]"
+labels: REFACT
 assignees: ''
 
 ---
