--- src/main/java/com/loopone/loopinbe/domain/loop/loop/loop
@@ -1,1 +1,0 @@
-..
