+++ src/main/java/com/loopone/loopinbe/domain/loop/loop/mainGoal/dto/req/MainGoalRequest.java
@@ -1,0 +1,18 @@
+package com.loopone.loopinbe.domain.loop.loop.mainGoal.dto.req;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDate;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class MainGoalRequest {
+    private String content;
+    private LocalDate deadline;
+    private Boolean checked;
+}
