+++ src/main/java/com/loopone/loopinbe/domain/loop/loop/mainGoal/dto/res/MainGoalWithSubGoalsResponse.java
@@ -1,0 +1,14 @@
+package com.loopone.loopinbe.domain.loop.loop.mainGoal.dto.res;
+
+import com.loopone.loopinbe.domain.loop.loop.subGoal.dto.res.SubGoalResponse;
+import lombok.Builder;
+import lombok.Data;
+
+import java.util.List;
+
+@Data
+@Builder
+public class MainGoalWithSubGoalsResponse {
+    private MainGoalResponse mainGoal;
+    private List<SubGoalResponse> subGoals;
+}
