--- src/main/java/com/loopone/loopinbe/domain/loop/teamLoop/mainGoal/dto/res/MainGoalWithSubGoalsResponse.java
@@ -1,14 +1,0 @@
-package com.loopone.loopinbe.domain.loop.teamLoop.mainGoal.dto.res;
-
-import com.loopone.loopinbe.domain.loop.teamLoop.subGoal.dto.res.SubGoalResponse;
-import lombok.Builder;
-import lombok.Data;
-
-import java.util.List;
-
-@Data
-@Builder
-public class MainGoalWithSubGoalsResponse {
-    private MainGoalResponse mainGoal;
-    private List<SubGoalResponse> subGoals;
-}
