--- src/main/java/com/loopone/loopinbe/domain/loop/teamLoop/mainGoal/entity/MainGoalPage.java
@@ -1,13 +1,0 @@
-package com.loopone.loopinbe.domain.loop.teamLoop.mainGoal.entity;
-
-import lombok.Data;
-import lombok.Getter;
-
-@Data
-public class MainGoalPage {
-    // 기본 page, size
-    private int page = 0;
-    private int size = 20;
-    @Getter
-    private static final int maxPageSize = 20;
-}
