--- src/main/java/com/loopone/loopinbe/domain/loop/teamLoop/subGoal/dto/res/SubGoalResponse.java
@@ -1,17 +1,0 @@
-package com.loopone.loopinbe.domain.loop.teamLoop.subGoal.dto.res;
-
-import lombok.Builder;
-import lombok.Data;
-
-import java.time.LocalDateTime;
-
-@Data
-@Builder
-public class SubGoalResponse {
-    private Long id;
-    private Long mainGoalId;
-    private String content;
-    private String dDay;
-    private Boolean checked;
-    private LocalDateTime createdAt;
-}
