--- src/main/java/com/loopone/loopinbe/domain/loop/teamLoop/subGoal/entity/SubGoal.java
@@ -1,36 +1,0 @@
-package com.loopone.loopinbe.domain.loop.teamLoop.subGoal.entity;
-
-import com.loopone.loopinbe.domain.account.member.entity.Member;
-import com.loopone.loopinbe.domain.loop.teamLoop.mainGoal.entity.MainGoal;
-import com.loopone.loopinbe.global.jpa.BaseEntity;
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import lombok.experimental.SuperBuilder;
-
-import java.time.LocalDate;
-
-@Entity
-@Setter
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-@SuperBuilder
-public class SubGoal extends BaseEntity {
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id", nullable = false)
-    private Member member;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "mainGoal_id", nullable = false)
-    private MainGoal mainGoal;
-
-    @Column(length = 500)
-    private String content;
-
-    private LocalDate deadline;
-
-    private Boolean checked;
-}
